@@ -4,17 +4,29 @@ build-backend = "setuptools.build_meta"
 
 [project]
 name = "effects-logging"
+description = "A composable logging framework built on algebraic effects"
+readme = "README.md"
+license = {text = "MIT"}
 dynamic = ["version"]
-requires-python = ">=3.10"
+requires-python = ">=3.12"
 dependencies = [
     "termcolor",
     "effects @ git+https://github.com/ludvb/effects",
 ]
 
+[project.urls]
+Homepage = "https://github.com/ludvb/effects-logging"
+Repository = "https://github.com/ludvb/effects-logging"
+Issues = "https://github.com/ludvb/effects-logging/issues"
+
 [project.optional-dependencies]
 dev = [
-    "pytest",
+    "pytest>=7.0",
+    "pytest-cov",
+    "pytest-xdist",
     "ruff",
+    "pyright",
+    "pre-commit",
 ]
 
 [tool.setuptools.dynamic]
@@ -24,3 +36,89 @@ version = {attr = "effects_logging.__version__.__version__"}
 minversion = "6.0"
 addopts = "--maxfail=1 --disable-warnings -q"
 testpaths = ["tests"]
+
+[tool.ruff]
+target-version = "py312"
+line-length = 100
+indent-width = 4
+exclude = [
+    ".git",
+    "__pycache__",
+    "build",
+    "dist",
+    "*.egg-info",
+    ".venv",
+    "venv",
+]
+
+[tool.ruff.lint]
+select = [
+    "E",    # pycodestyle errors
+    "W",    # pycodestyle warnings
+    "F",    # pyflakes
+    "I",    # isort
+    "B",    # flake8-bugbear
+    "C4",   # flake8-comprehensions
+    "UP",   # pyupgrade
+    "ARG",  # flake8-unused-arguments
+    "SIM",  # flake8-simplify
+    "TID",  # flake8-tidy-imports
+    "RUF",  # Ruff-specific rules
+]
+ignore = [
+    "E501",   # Line too long (handled by formatter)
+    "ARG001", # Unused function argument
+    "ARG002", # Unused method argument
+    "B008",   # Do not perform function calls in argument defaults
+    "SIM108", # Use ternary operator
+    "SIM117", # Nested with statements
+    "UP035",  # Deprecated typing imports
+    "UP047",  # Use type parameters
+]
+
+[tool.ruff.lint.per-file-ignores]
+"__init__.py" = ["F401", "F403"]
+"test_*.py" = ["ARG"]
+
+[tool.ruff.lint.isort]
+known-first-party = ["effects_logging", "effects"]
+combine-as-imports = true
+
+[tool.ruff.format]
+quote-style = "double"
+indent-style = "space"
+skip-magic-trailing-comma = false
+line-ending = "auto"
+
+[tool.pyright]
+include = ["src", "tests"]
+exclude = ["**/__pycache__", "build", "dist", "*.egg-info"]
+pythonVersion = "3.12"
+pythonPlatform = "All"
+typeCheckingMode = "standard"
+
+useLibraryCodeForTypes = true
+reportMissingTypeStubs = false
+reportMissingImports = true
+
+# Allow gradual typing
+reportMissingParameterType = "none"
+reportUnknownParameterType = "none"
+reportMissingTypeArgument = "none"
+reportUnknownMemberType = "none"
+reportUnknownArgumentType = "none"
+reportUnknownLambdaType = "none"
+reportUnknownVariableType = "none"
+
+# Catch real problems
+reportMatchNotExhaustive = "error"
+reportUnusedVariable = "warning"
+reportUnusedImport = "warning"
+reportUnusedClass = "warning"
+reportUnusedFunction = "warning"
+reportPrivateUsage = "warning"
+reportDuplicateImport = "warning"
+
+# Effects library patterns
+reportIncompatibleMethodOverride = "none"
+reportIncompatibleVariableOverride = "none"
